Remove the parent Snyk organization parameter from the quick start.
Passing a parent organization from which to copy settings for a new
organization within Snyk does not actually copy the Members of the
source organization to the new. This was essentially the only reason for
including it, so now we can save the user an input and make the docs a
bit shorter.